# src/Dreamlines.Presentation.Api/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Dreamlines Bookings Api",
    "description" : "This is prototype of the Dreamlines Bookings API\nThe api allows access the bookings and their related SalesUnits.\n",
    "termsOfService" : "http://swagger.io/terms/",
    "contact" : {
      "email" : "apiteam@swagger.io"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/DreamlinesApi/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "https://petstore.swagger.io/v2"
  } ],
  "tags" : [ {
    "name" : "booking",
    "description" : "Operations related to Booking entity."
  }, {
    "name" : "salesunit",
    "description" : "Operations related to Sales Unit entity."
  } ],
  "paths" : {
    "/booking/findBySalesUnitId/{salesUnitId}" : {
      "get" : {
        "tags" : [ "booking" ],
        "summary" : "Returns all the bookings that belong to a salesunit with t",
        "description" : "TODO",
        "operationId" : "findBySalesUnitId",
        "parameters" : [ {
          "name" : "salesUnitId",
          "in" : "path",
          "description" : "ID of the salesUnit",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Booking"
                  },
                  "x-content-type" : "application/json"
                }
              },
              "application/xml" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Booking"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid salesunit id"
          }
        }
      }
    },
    "/booking/{bookingId}" : {
      "get" : {
        "tags" : [ "booking" ],
        "summary" : "Finds a booking by ID",
        "description" : "Returns a single booking",
        "operationId" : "getBookingById",
        "parameters" : [ {
          "name" : "bookingId",
          "in" : "path",
          "description" : "ID of booking to return",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Booking"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Booking"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid ID supplied"
          },
          "404" : {
            "description" : "Booking not found"
          }
        }
      }
    },
    "/salesunit/search" : {
      "get" : {
        "tags" : [ "salesunit" ],
        "summary" : "Lists the salesunits",
        "description" : "Returns a list of Sales Units, showing the name and the total price of their related Bookings. Results can be filtered by bookingIds or ship names.",
        "operationId" : "salesunitsearch",
        "parameters" : [ {
          "name" : "searchTerm",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "description" : "Autocomplete search term that accepts either Booking Ids or Ship Names"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "description" : "The number of items to skip before starting to collect the result set"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "description" : "The numbers of items to return"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SalesUnitSummary"
                  },
                  "x-content-type" : "application/json"
                }
              },
              "application/xml" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SalesUnitSummary"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Booking" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "shipId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "bookingDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "price" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "example" : {
          "price" : 1,
          "bookingDate" : "2000-01-23T04:56:07.000+00:00",
          "id" : 0,
          "shipId" : 6
        },
        "xml" : {
          "name" : "Booking"
        }
      },
      "SalesUnitSummary" : {
        "type" : "object",
        "properties" : {
          "bookingPriceSum" : {
            "type" : "integer",
            "format" : "int64"
          },
          "salesUnit" : {
            "$ref" : "#/components/schemas/SalesUnit"
          }
        },
        "example" : {
          "salesUnit" : {
            "country" : "country",
            "bookingPriceSummary" : 1,
            "name" : "name",
            "currency" : "currency",
            "id" : 6
          },
          "bookingPriceSum" : 0
        },
        "xml" : {
          "name" : "SalesUnitSummary"
        }
      },
      "SalesUnit" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "country" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "bookingPriceSummary" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "example" : {
          "country" : "country",
          "bookingPriceSummary" : 1,
          "name" : "name",
          "currency" : "currency",
          "id" : 6
        },
        "xml" : {
          "name" : "SalesUnit"
        }
      }
    }
  }
}